fix: adapt repositories to new id formats

+++ src/main/kotlin/es/uib/tfg/sports_league_backend/phase/infrastructure/repository/PhaseRepository.kt
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface PhaseRepository: JpaRepository<Phase, Int> {
-    fun findAllByLeagueId(leagueId: Int): List<Phase>
+interface PhaseRepository: JpaRepository<Phase, Long> {
+    fun findAllByLeagueId(leagueId: Long): List<Phase>
 }
